requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
cors==1.0.1
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gensim==4.3.3
gevent==24.10.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
kombu==5.4.2
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.57.0
pandas==2.2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.5.2
scipy==1.13.1
setuptools==75.3.0
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tldextract==5.1.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.0
zope.event==5.0
zope.interface==7.1.1
